+++ packages/components/organisms/f-checkout/stories/Checkout.stories.js
@@ -42,6 +42,7 @@ const getCheckoutAccessForbiddenUrl = '/checkout-403-get-error.json';
 const getCheckoutErrorUrl = '/checkout-500-get-error.json';
 const paymentPageUrlPrefix = '#/pay'; // Adding the "#" so we don't get redirect out of the component in Storybook
 const getGeoLocationUrl = '/get-geo-location.json';
+const getCustomerUrl = '/get-customer.json';
 
 CheckoutMock.setupCheckoutMethod(getCheckoutDeliveryUrl);
 CheckoutMock.setupCheckoutMethod(getCheckoutDeliveryAsapUrl);
@@ -68,6 +69,8 @@ CheckoutMock.setupCheckoutMethod(placeOrderDuplicateUrl);
 CheckoutMock.setupCheckoutMethod(getCheckoutAccessForbiddenUrl);
 CheckoutMock.setupCheckoutMethod(getCheckoutErrorUrl);
 CheckoutMock.setupCheckoutMethod(getGeoLocationUrl);
+CheckoutMock.setupCheckoutMethod(getCustomerUrl);
+
 CheckoutMock.passThroughAny();
 
 const restraurantNotTakingOrders = 'Restaurant Not Taking Orders Issue (Response from server but order not fulfillable)';
@@ -128,7 +131,8 @@ export const CheckoutComponent = () => ({
             getAddressUrl,
             loginUrl: '/login',
             paymentPageUrlPrefix,
-            getGeoLocationUrl
+            getGeoLocationUrl,
+            getCustomerUrl
         };
     },
     props: {
@@ -227,6 +231,7 @@ export const CheckoutComponent = () => ({
         ':paymentPageUrlPrefix="paymentPageUrlPrefix" ' +
         'applicationName="Storybook" ' +
         ':getGeoLocationUrl="getGeoLocationUrl" ' +
+        ':getCustomerUrl="getCustomerUrl" ' +
         // eslint-disable-next-line no-template-curly-in-string
         ' :key="`${locale},${getCheckoutUrl},${updateCheckoutUrl},${checkoutAvailableFulfilmentUrl},${authToken},${createGuestUrl},${getBasketUrl},${getAddressUrl},${placeOrderUrl},${paymentPageUrlPrefix},${getGeoLocationUrl}`" />'
 });