+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -3,6 +3,23 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v0.153.0
+------------------------------
+*July 6, 2021*
+
+### Added
+- Chosen time to eventData to be captured by the error log.
+
+
+v0.152.0
+------------------------------
+*July 6, 2021*
+
+### Fixed
+- Returning guest user's mobile phone number not populating
+
+
 v0.151.0
 ------------------------------
 *July 5, 2021*